+++ genesis-vocali/modules/auth/login-service.ts
@@ -1,15 +1,15 @@
 import supabase from '@/services/auth/auth-supabase.service';
 
-export async function registerUser(email, password) {
+export async function loginUser(email, password) {
   try {
-    const { user, error } = await supabase.auth.signUp({
+    const { user, error } = await supabase.auth.signIn({
       email,
       password,
     });
     if (error) throw error;
     return user;
   } catch (error) {
-    console.error('Registration error:', error);
+    console.error('Login error:', error);
     throw error;
   }
 }
