+++ genesis-vocali/modules/auth/register-service.spec.ts
@@ -2,43 +2,22 @@
  * @jest-environment node
  */
 
-import { supabase } from "@/services/auth/auth-supabase.service";
+import supabase from "@/services/auth/auth-supabase.service";
 import { registerUser } from "./register-service";
 
 describe("registerService (Integration Test)", () => {
-  const validEmail = "test@example.com";
-  const validPassword = "strong_password";
+  it("registers a user with valid email and password", async () => {
+    const newUserEmail =  `newuser${Date.now()}@example.com`;
+    const newUserPassword = "another_strong_password";
 
-  beforeEach(async () => {
-    const { data, error } = await supabase
-      .from("users")
-      .delete()
-      .match({ email: validEmail });
-    if (error) throw error;
-  });
-
-  afterEach(async () => {
-    const { data, error } = await supabase
-      .from("users")
-      .delete()
-      .match({ email: validEmail });
-    if (error) throw error;
+    expect(async () => await registerUser(newUserEmail, newUserPassword)).not.toThrow();
   });
 
   it("registers a user with valid email and password", async () => {
-    const newUserEmail = `newuser${Date.now()}@example.com`;
+    const newUserEmail = 'mrtnomar15@gmail.com'; // `newuser${Date.now()}@example.com`;
     const newUserPassword = "another_strong_password";
 
     const registeredUser = await registerUser(newUserEmail, newUserPassword);
-    console.log(registeredUser);
     expect(registeredUser).not.toBeNull();
-
-    const { data, error } = await supabase
-      .from("users")
-      .select("*")
-      .single()
-      .eq("email", newUserEmail);
-    expect(error).toBeNull();
-    expect(data).toHaveProperty("email", newUserEmail);
   });
 });
