+++ genesis-vocali/services/auth/auth-supabase.service.ts
@@ -1,7 +1,7 @@
-import { createClient } from '@supabase/js';
-import config from '@/resources/config';
+import { createClient } from "@supabase/supabase-js";
+import config from "@/resources/utils/config";
 
 const supabaseUrl = config.SUPABASE_URL;
 const supabaseAnonKey = config.SUPABASE_API_KEY;
 
-export const supabase = createClient(supabaseUrl, supabaseAnonKey);
+export default createClient(supabaseUrl, supabaseAnonKey);
